# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GhostTech API",
    "description": "Бизнес-задачи, рейтинг, команды, отклики и этапы работы.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/tasks/analyze": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Analyze",
        "operationId": "analyze_api_tasks_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/build-card": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Build Card",
        "operationId": "build_card_api_tasks_build_card_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalyzeRead": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 3,
            "title": "Questions"
          },
          "filled_fields": {
            "items": {
              "type": "string",
              "enum": [
                "title",
                "topic",
                "context",
                "need",
                "users",
                "data",
                "constraints",
                "expected_result",
                "success_criteria",
                "contact",
                "interaction_format"
              ]
            },
            "type": "array",
            "maxItems": 11,
            "title": "Filled Fields"
          },
          "missing_fields": {
            "items": {
              "type": "string",
              "enum": [
                "title",
                "topic",
                "context",
                "need",
                "users",
                "data",
                "constraints",
                "expected_result",
                "success_criteria",
                "contact",
                "interaction_format"
              ]
            },
            "type": "array",
            "maxItems": 11,
            "title": "Missing Fields"
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "openai"
            ],
            "title": "Mode",
            "default": "demo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "questions",
          "filled_fields",
          "missing_fields"
        ],
        "title": "AnalyzeRead"
      },
      "AnalyzeRequest": {
        "properties": {
          "draft_text": {
            "type": "string",
            "maxLength": 20000,
            "title": "Draft Text"
          },
          "topic": {
            "type": "string",
            "maxLength": 100,
            "title": "Topic",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "draft_text"
        ],
        "title": "AnalyzeRequest"
      },
      "Answer": {
        "properties": {
          "question_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Question Id"
          },
          "field": {
            "type": "string",
            "enum": [
              "title",
              "topic",
              "context",
              "need",
              "users",
              "data",
              "constraints",
              "expected_result",
              "success_criteria",
              "contact",
              "interaction_format"
            ],
            "title": "Field"
          },
          "answer": {
            "type": "string",
            "maxLength": 5000,
            "title": "Answer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question_id",
          "field",
          "answer"
        ],
        "title": "Answer"
      },
...
      "BuildRead": {
        "properties": {
          "card": {
            "$ref": "#/components/schemas/Card"
          },
          "warnings": {
            "items": {
              "type": "string",
              "maxLength": 1000,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 20,
            "title": "Warnings"
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "openai"
            ],
            "title": "Mode",
            "default": "demo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "card",
          "warnings"
        ],
        "title": "BuildRead"
      },
      "BuildRequest": {
        "properties": {
          "draft_text": {
            "type": "string",
            "maxLength": 20000,
            "title": "Draft Text"
          },
          "topic": {
            "type": "string",
            "maxLength": 100,
            "title": "Topic",
            "default": ""
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/Answer"
            },
            "type": "array",
            "maxItems": 11,
            "title": "Answers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "draft_text"
        ],
        "title": "BuildRequest"
      },
      "Card": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title",
            "default": ""
          },
          "topic": {
            "type": "string",
            "maxLength": 100,
            "title": "Topic",
            "default": ""
          },
          "context": {
            "type": "string",
            "maxLength": 5000,
            "title": "Context",
            "default": ""
          },
          "need": {
            "type": "string",
            "maxLength": 5000,
            "title": "Need",
            "default": ""
          },
          "users": {
            "type": "string",
            "maxLength": 5000,
            "title": "Users",
            "default": ""
          },
          "data": {
            "type": "string",
            "maxLength": 5000,
            "title": "Data",
            "default": ""
          },
          "constraints": {
            "type": "string",
            "maxLength": 5000,
            "title": "Constraints",
            "default": ""
          },
          "expected_result": {
            "type": "string",
            "maxLength": 5000,
            "title": "Expected Result",
            "default": ""
          },
          "success_criteria": {
            "type": "string",
            "maxLength": 5000,
            "title": "Success Criteria",
            "default": ""
          },
          "contact": {
            "type": "string",
            "maxLength": 500,
            "title": "Contact",
            "default": ""
          },
          "interaction_format": {
            "type": "string",
            "maxLength": 5000,
            "title": "Interaction Format",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Card"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Question": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Id"
          },
          "field": {
            "type": "string",
            "enum": [
              "title",
              "topic",
              "context",
              "need",
              "users",
              "data",
              "constraints",
              "expected_result",
              "success_criteria",
              "contact",
              "interaction_format"
            ],
            "title": "Field"
          },
          "question": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "field",
          "question"
        ],
        "title": "Question"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}